+++ brownford/src/main/java/com/brownford/repository/FacultyRepository.java
@@ -7,5 +7,6 @@ import java.util.Optional;
 
 public interface FacultyRepository extends JpaRepository<Faculty, Long> {
     Optional<Faculty> findByFacultyId(String facultyId);
+
     Optional<Faculty> findByUser(User user);
 }